+++ Source/Common/IGListExperiments.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016-present, Facebook, Inc.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -24,6 +24,10 @@ typedef NS_OPTIONS (NSInteger, IGListExperiment) {
     IGListExperimentDeferredToObjectCreation = 1 << 6,
     /// Test getting collection view at update time.
     IGListExperimentGetCollectionViewAtUpdate = 1 << 7,
+    /// Test skipping layout when UICollectionView isn't visible
+    IGListExperimentSkipLayout = 1 << 8,
+    /// Test updater diffing performed on a serial background queue.
+    IGListExperimentBackgroundDiffingSerial = 1 << 9,
 };
 
 /**