+++ Source/IGListCollectionViewLayout.mm
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016-present, Facebook, Inc.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -415,13 +415,17 @@ static void adjustZIndexForAttributes(UICollectionViewLayoutAttributes *attribut
 - (BOOL)shouldInvalidateLayoutForBoundsChange:(CGRect)newBounds {
     const CGRect oldBounds = self.collectionView.bounds;
 
+    // always invalidate for size changes
+    if (!CGSizeEqualToSize(oldBounds.size, newBounds.size)) {
+        return YES;
+    }
+
     // if the y origin has changed, only invalidate when using sticky headers
     if (CGRectGetMinInDirection(newBounds, self.scrollDirection) != CGRectGetMinInDirection(oldBounds, self.scrollDirection)) {
         return self.stickyHeaders;
     }
 
-    // always invalidate for size changes
-    return !CGSizeEqualToSize(oldBounds.size, newBounds.size);
+    return NO;
 }
 
 - (void)prepareLayout {