+++ Tests/IGListAdapterUpdaterTests.m
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016-present, Facebook, Inc.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -17,6 +17,7 @@
 
 #define genExpectation [self expectationWithDescription:NSStringFromSelector(_cmd)]
 #define waitExpectation [self waitForExpectationsWithTimeout:30 handler:nil]
+#define genToBlock ^NSArray *{ return to; }
 
 @interface IGListAdapterUpdaterTests : XCTestCase
 
@@ -94,6 +95,17 @@
     self.dataSource.sections = @[];
     [self.updater performReloadDataWithCollectionViewBlock:[self collectionViewBlock]];
     XCTAssertEqual([self.collectionView numberOfSections], 0);
+}
+
+- (void)test_whenReloadingDataWithNilDataSourceBefore_thatCollectionViewNotCrash {
+    self.dataSource.sections = @[[IGSectionObject sectionWithObjects:@[@1]], [IGSectionObject sectionWithObjects:@[@2]]];
+    [self.updater performReloadDataWithCollectionViewBlock:[self collectionViewBlock]];
+    XCTAssertEqual([self.collectionView numberOfSections], 2);
+    
+    self.collectionView.dataSource = nil;
+    self.dataSource.sections = @[];
+    [self.updater performReloadDataWithCollectionViewBlock:[self collectionViewBlock]];
+    XCTAssertEqual([self.collectionView numberOfSections], 1); // Setting collectionView's dataSource to nil would yield a single section by default.
 }
 
 - (void)test_whenInsertingSection_thatCollectionViewUpdates {
@@ -435,7 +447,7 @@
     id mockDelegate = [OCMockObject mockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     id compilerFriendlyNil = nil;
-    [[mockDelegate reject] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:compilerFriendlyNil];
+    [[mockDelegate reject] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:compilerFriendlyNil fromObjects:@[] toObjects:@[] listIndexSetResult:OCMOCK_ANY];
     [[mockDelegate reject] listAdapterUpdater:self.updater didPerformBatchUpdates:[OCMArg any] collectionView:compilerFriendlyNil];
     [self.updater performBatchUpdatesWithCollectionViewBlock:^UICollectionView *{ return compilerFriendlyNil; }];
     [mockDelegate verify];
@@ -489,16 +501,14 @@
     id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[]]];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:self.dataSource.sections toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        return result.inserts.firstIndex == 0 && result.moves.count == 0 && result.updates.count == 0 && result.deletes.count == 0;
+    }]];
     [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:OCMOCK_ANY collectionView:self.collectionView];
 
     XCTestExpectation *expectation = genExpectation;
-    IGListToObjectBlock to = ^NSArray *{
-        return @[
-                 [IGSectionObject sectionWithObjects:@[]]
-                 ];
-    };
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:self.dataSource.sections toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:self.dataSource.sections toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;
@@ -516,7 +526,7 @@
     // in the future, but we configure the mock delegate to allow any call
     // except the batch updates calls.
 
-    [[mockDelegate reject] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    [[mockDelegate reject] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:@[] toObjects:@[] listIndexSetResult:OCMOCK_ANY];
     [[mockDelegate reject] listAdapterUpdater:self.updater didPerformBatchUpdates:OCMOCK_ANY collectionView:self.collectionView];
 
     XCTestExpectation *expectation = genExpectation;
@@ -525,6 +535,28 @@
                  [IGSectionObject sectionWithObjects:@[]]
                  ];
     };
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:self.dataSource.sections toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+        [expectation fulfill];
+    }];
+    waitExpectation;
+    [mockDelegate verify];
+}
+
+- (void)test_whenCollectionViewNotInWindow_andBackgroundReloadFlag_isDefaultYES_andDataSourceWasSetToNilBefore_thatCollectionViewNotCrash {
+    [self.collectionView removeFromSuperview];
+    
+    id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
+    self.updater.delegate = mockDelegate;
+    [[mockDelegate reject] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:@[] toObjects:@[] listIndexSetResult:OCMOCK_ANY];
+    [[mockDelegate reject] listAdapterUpdater:self.updater didPerformBatchUpdates:OCMOCK_ANY collectionView:self.collectionView];
+    
+    XCTestExpectation *expectation = genExpectation;
+    IGListToObjectBlock to = ^NSArray *{
+        return @[
+                 [IGSectionObject sectionWithObjects:@[]]
+                 ];
+    };
+    self.collectionView.dataSource = nil;
     [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:self.dataSource.sections toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
@@ -668,22 +700,51 @@
                                                                                           deleteIndexPaths:@[]
                                                                                           updateIndexPaths:@[]
                                                                                             moveIndexPaths:@[]];
+    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id"]];
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@2] identifier:@"id"]];
+    self.dataSource.sections = from;
 
     id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        return result.inserts.count == 0 && result.deletes.count == 0 && result.moves.count == 0 && result.updates.firstIndex == 0;
+    }]];
     [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
     
     XCTestExpectation *expectation = genExpectation;
-    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id"]];
-    IGListToObjectBlock to = ^NSArray *{
-        // Update the items
-        return @[[IGSectionObject sectionWithObjects:@[@2] identifier:@"id"]];
-    };
+
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+        [expectation fulfill];
+    }];
+    waitExpectation;
+    [mockDelegate verify];
+}
+
+- (void)test_whenPerformUpdates_dataSourceWasSetToNil_shouldNotCrash {
+    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"]];
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@2] identifier:@"id1"],
+                                       [IGSectionObject sectionWithObjects:@[@22] identifier:@"id2"]];
     self.dataSource.sections = from;
+    id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
+    self.updater.delegate = mockDelegate;
+    [mockDelegate setExpectationOrderMatters:YES];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        if (result.deletes.count != 0 || result.moves.count != 0) {
+            return NO;
+        }
+        // Make sure we note that index 1 is updated (id1 from @[@1] -> @[@2]), and "id2" was inserted at index 1
+        return result.updates.firstIndex == 0 && result.inserts.firstIndex == 1;
+    }]];
+    [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:OCMOCK_ANY collectionView:self.collectionView];
     
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    XCTestExpectation *expectation = genExpectation;
+    
+    // Manually set the data source to be nil.
+    self->_collectionView.dataSource = nil;
+    
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:^(NSArray * _Nonnull toObjects) {
+    }  completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;
@@ -702,22 +763,21 @@
                                                                                           deleteIndexPaths:@[]
                                                                                           updateIndexPaths:@[[NSIndexPath indexPathForItem:0 inSection:0]]
                                                                                             moveIndexPaths:@[]];
-    
+    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id"]];
+    // Update the items
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@2] identifier:@"id"]];
+    self.dataSource.sections = from;
     id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        return result.inserts.count == 0 && result.deletes.count == 0 && result.moves.count == 0 && result.updates.firstIndex == 0;
+    }]];
     [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
     
     XCTestExpectation *expectation = genExpectation;
-    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id"]];
-    IGListToObjectBlock to = ^NSArray *{
-        // Update the items
-        return @[[IGSectionObject sectionWithObjects:@[@2] identifier:@"id"]];
-    };
-    self.dataSource.sections = from;
     
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;
@@ -734,22 +794,22 @@
                                                                                           deleteIndexPaths:@[]
                                                                                           updateIndexPaths:@[]
                                                                                             moveIndexPaths:@[]];
+    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id"]];
+    // more items in the section
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@1, @2] identifier:@"id"]];
+    self.dataSource.sections = from;
     
     id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        return result.inserts.count == 0 && result.deletes.count == 0 && result.moves.count == 0 && result.updates.firstIndex == 0;
+    }]];
     [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
     
     XCTestExpectation *expectation = genExpectation;
-    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id"]];
-    IGListToObjectBlock to = ^NSArray *{
-        // more items in the section
-        return @[[IGSectionObject sectionWithObjects:@[@1, @2] identifier:@"id"]];
-    };
-    self.dataSource.sections = from;
     
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;
@@ -766,23 +826,28 @@
                                                                                           deleteIndexPaths:@[]
                                                                                           updateIndexPaths:@[]
                                                                                             moveIndexPaths:@[]];
+    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"],
+                                         [IGSectionObject sectionWithObjects:@[@2] identifier:@"id2"]];
+    // move section, and also update the item for "id2"
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@22] identifier:@"id2"],
+                                       [IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"]];
+    self.dataSource.sections = from;
+
     id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        if (result.inserts.count != 0 || result.deletes.count != 0) {
+            return NO;
+        }
+        // Make sure we note that index 1 is updated (id2 from @[@2] -> @[@22], "id1" moved from section 0 -> 1, and "id2" moved from section 1 -> 0
+        return result.updates.firstIndex == 1 && result.moves.count == 2 && [result.moves containsObject:[[IGListMoveIndex alloc] initWithFrom:0 to:1]] && [result.moves containsObject:[[IGListMoveIndex alloc] initWithFrom:1 to:0]];
+    }]];
     [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
     
     XCTestExpectation *expectation = genExpectation;
-    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"],
-                                         [IGSectionObject sectionWithObjects:@[@2] identifier:@"id2"]];
-    IGListToObjectBlock to = ^NSArray *{
-        // move section, and also update the item for "id2"
-        return @[[IGSectionObject sectionWithObjects:@[@22] identifier:@"id2"],
-                 [IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"]];
-    };
-    self.dataSource.sections = from;
     
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;
@@ -799,23 +864,28 @@
                                                                                           deleteIndexPaths:@[]
                                                                                           updateIndexPaths:@[]
                                                                                             moveIndexPaths:@[]];
+    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1, @2, @3] identifier:@"id1"],
+                                         [IGSectionObject sectionWithObjects:@[@2] identifier:@"id2"]];
+    // move section, and also update the item for "id2"
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@22] identifier:@"id2"],
+                                       [IGSectionObject sectionWithObjects:@[@1, @2, @3] identifier:@"id1"]];
+    self.dataSource.sections = from;
+
     id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        if (result.inserts.count != 0 || result.deletes.count != 0) {
+            return NO;
+        }
+        // Make sure we note that index 1 is updated (id2 from @[@2] -> @[@22], "id1" moved from section 0 -> 1, and "id2" moved from section 1 -> 0
+        return result.updates.firstIndex == 1 && result.moves.count == 2 && [result.moves containsObject:[[IGListMoveIndex alloc] initWithFrom:0 to:1]] && [result.moves containsObject:[[IGListMoveIndex alloc] initWithFrom:1 to:0]];
+    }]];
     [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
     
     XCTestExpectation *expectation = genExpectation;
-    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1, @2, @3] identifier:@"id1"],
-                                         [IGSectionObject sectionWithObjects:@[@2] identifier:@"id2"]];
-    IGListToObjectBlock to = ^NSArray *{
-        // move section, and also update the item for "id2"
-        return @[[IGSectionObject sectionWithObjects:@[@22] identifier:@"id2"],
-                 [IGSectionObject sectionWithObjects:@[@1, @2, @3] identifier:@"id1"]];
-    };
-    self.dataSource.sections = from;
     
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;
@@ -833,57 +903,64 @@
                                                                                           deleteIndexPaths:@[]
                                                                                           updateIndexPaths:@[]
                                                                                             moveIndexPaths:@[]];
-    id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
-    self.updater.delegate = mockDelegate;
-    [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
-    [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
-    
-    XCTestExpectation *expectation = genExpectation;
     NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"],
                                          [IGSectionObject sectionWithObjects:@[@2] identifier:@"id2"],
                                          [IGSectionObject sectionWithObjects:@[@3] identifier:@"id3"]];
-    IGListToObjectBlock to = ^NSArray *{
-        // move section, and also update the items for "id2"
-        return @[[IGSectionObject sectionWithObjects:@[@22, @23] identifier:@"id2"],
-                 [IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"],
-                 [IGSectionObject sectionWithObjects:@[@3] identifier:@"id3"]];
-    };
+    // move section, and also update the items for "id2"
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@22, @23] identifier:@"id2"],
+                                       [IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"],
+                                       [IGSectionObject sectionWithObjects:@[@3] identifier:@"id3"]];
     self.dataSource.sections = from;
+    id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
+    self.updater.delegate = mockDelegate;
+    [mockDelegate setExpectationOrderMatters:YES];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        if (result.inserts.count != 0 || result.deletes.count != 0) {
+            return NO;
+        }
+        // Make sure we note that index 1 is updated (id2 from @[@2] -> @[@22], "id1" moved from section 0 -> 1, "id2" moved from section 1 -> 0
+        return result.updates.firstIndex == 1 && result.moves.count == 2 && [result.moves containsObject:[[IGListMoveIndex alloc] initWithFrom:0 to:1]] && [result.moves containsObject:[[IGListMoveIndex alloc] initWithFrom:1 to:0]];
+    }]];
+    [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
     
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    XCTestExpectation *expectation = genExpectation;
+    
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;
     [mockDelegate verify];
 }
 
-- (void)test_whenReloadIsCalledWithSectionInsertAndUpdate_andPreferItemReload_orderHappened {
+- (void)test_whenReloadIsCalledWithSectionInsertAndUpdate_andPreferItemReload_noItemReloads {
     self.updater.preferItemReloadsForSectionReloads = YES;
     
-    IGListBatchUpdateData *expectedBatchUpdateData = [[IGListBatchUpdateData alloc] initWithInsertSections:[NSIndexSet indexSetWithIndex:1]
-                                                                                            deleteSections:[NSIndexSet new]
+    IGListBatchUpdateData *expectedBatchUpdateData = [[IGListBatchUpdateData alloc] initWithInsertSections:[NSIndexSet indexSetWithIndexesInRange:NSMakeRange(0, 2)]
+                                                                                            deleteSections:[NSIndexSet indexSetWithIndex:0]
                                                                                               moveSections:[NSSet new]
                                                                                           insertIndexPaths:@[]
                                                                                           deleteIndexPaths:@[]
-                                                                                          updateIndexPaths:@[[NSIndexPath indexPathForItem:0 inSection:0]]
+                                                                                          updateIndexPaths:@[]
                                                                                             moveIndexPaths:@[]];
+    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"]];
+    NSArray<IGSectionObject *> *to = @[[IGSectionObject sectionWithObjects:@[@2] identifier:@"id1"],
+                                       [IGSectionObject sectionWithObjects:@[@22] identifier:@"id2"]];
+    self.dataSource.sections = from;
     id mockDelegate = [OCMockObject niceMockForProtocol:@protocol(IGListAdapterUpdaterDelegate)];
     self.updater.delegate = mockDelegate;
     [mockDelegate setExpectationOrderMatters:YES];
-    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView];
+    [[mockDelegate expect] listAdapterUpdater:self.updater willPerformBatchUpdatesWithCollectionView:self.collectionView fromObjects:from toObjects:to listIndexSetResult:[OCMArg checkWithBlock:^BOOL(IGListIndexSetResult *result) {
+        if (result.deletes.count != 0 || result.moves.count != 0) {
+            return NO;
+        }
+        // Make sure we note that index 1 is updated (id1 from @[@1] -> @[@2]), and "id2" was inserted at index 1
+        return result.updates.firstIndex == 0 && result.inserts.firstIndex == 1;
+    }]];
     [[mockDelegate expect] listAdapterUpdater:self.updater didPerformBatchUpdates:expectedBatchUpdateData collectionView:self.collectionView];
     
     XCTestExpectation *expectation = genExpectation;
-    NSArray<IGSectionObject *> *from = @[[IGSectionObject sectionWithObjects:@[@1] identifier:@"id1"]];
-    IGListToObjectBlock to = ^NSArray *{
-        // insert new section id2, and also update for section id1
-        return @[[IGSectionObject sectionWithObjects:@[@2] identifier:@"id1"],
-                 [IGSectionObject sectionWithObjects:@[@22] identifier:@"id2"]];
-    };
-    self.dataSource.sections = from;
     
-    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:to animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
+    [self.updater performUpdateWithCollectionViewBlock:[self collectionViewBlock] fromObjects:from toObjectsBlock:genToBlock animated:NO objectTransitionBlock:self.updateBlock completion:^(BOOL finished) {
         [expectation fulfill];
     }];
     waitExpectation;